simplify interface
