+++ src/index.js
@@ -1,5 +1,4 @@
 import filter from './filter';
-import transchoice from './transchoice';
 
 /* eslint no-console: ["error", { allow: ["warn", "error", "no-undef"] }] */
 const install = (Vue) => {
@@ -7,13 +6,8 @@ const install = (Vue) => {
     console.warn('[filter duplication]: There is already a filter named `trans` registered');
     return;
   }
-  if (Vue.filter('transchoice')) {
-    console.warn('[filter duplication]: There is already a filter named `transchoice` registered');
-    return;
-  }
 
   Vue.filter('trans', filter);
-  Vue.filter('transchoice', transchoice);
 };
 
 /* global Vue */
